+++ Libs/robotkit/quat.h
@@ -5,15 +5,15 @@
 #include "matrix.h"
 
 typedef struct {
-    float w, x, y, z;
+    double w, x, y, z;
 } quaternion_t;
 
 typedef struct {
-    float roll, pitch, yaw;
+    double roll, pitch, yaw;
 } euler_angle_t;
 
 // Initialize quaternion
-void quat_init(quaternion_t* q, float w, float x, float y, float z);
+void quat_init(quaternion_t* q, double w, double x, double y, double z);
 
 // Set quaternion
 void quat_set(quaternion_t* q, const quaternion_t* q_new);
@@ -34,10 +34,10 @@ void quat_mult(quaternion_t* result, const quaternion_t* a, const quaternion_t* 
 void quat_conjugate(quaternion_t* result, const quaternion_t* q);
 
 // Create quaternion from axis-angle
-void quat_from_axis_angle(quaternion_t* q, const vector3d_t* axis, float angle);
+void quat_from_axis_angle(quaternion_t* q, const vector3d_t* axis, double angle);
 
 // Quaternion from rotations
-void quat_from_euler(quaternion_t *result, float roll, float pitch, float yaw);
+void quat_from_euler(quaternion_t *result, double roll, double pitch, double yaw);
 
 // Convert quaternion to Euler angles (ZYX order) and store result in output parameter
 void quat_to_euler(euler_angle_t* euler, const quaternion_t* q);
@@ -49,9 +49,9 @@ void quat_to_rot_matrix(matrix_t* rot, const quaternion_t* q);
 void quat_rotate_vector(vector3d_t* result, const quaternion_t* q, const vector3d_t* v);
 
 // Quaternion exponential map (for small angles)
-void quat_exp_map(quaternion_t* result, const vector3d_t* w, float dt);
+void quat_exp_map(quaternion_t* result, const vector3d_t* w, double dt);
 
 // Measurement function h(q): Predicts accelerometer reading from quaternion state
-void quat_to_accel(vector3d_t *a, const quaternion_t *q, float g);
+void quat_to_accel(vector3d_t *a, const quaternion_t *q, double g);
 
 #endif
